+++ database/migrations/2024_11_14_173959_add_foreign_keys_to_reserva_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('reserva', function (Blueprint $table) {
-            $table->foreign(['cliente_id_cliente'], 'fk_reserva_cliente')->references(['id_cliente'])->on('cliente')->onUpdate('no action')->onDelete('no action');
-            $table->foreign(['deposito_id_deposito'], 'fk_reserva_deposito')->references(['id_deposito'])->on('deposito')->onUpdate('no action')->onDelete('no action');
-            $table->foreign(['tour_id_tour'], 'fk_reserva_tour')->references(['id_tour'])->on('tour')->onUpdate('no action')->onDelete('no action');
-            $table->foreign(['usuario_id_usuario'], 'fk_reserva_usuario')->references(['id_usuario'])->on('usuario')->onUpdate('no action')->onDelete('no action');
-        });
+        if (!Schema::hasTable('reserva')) {
+            Schema::table('reserva', function (Blueprint $table) {
+                $table->foreign(['cliente_id_cliente'], 'fk_reserva_cliente')->references(['id_cliente'])->on('cliente')->onUpdate('no action')->onDelete('no action');
+                $table->foreign(['deposito_id_deposito'], 'fk_reserva_deposito')->references(['id_deposito'])->on('deposito')->onUpdate('no action')->onDelete('no action');
+                $table->foreign(['tour_id_tour'], 'fk_reserva_tour')->references(['id_tour'])->on('tour')->onUpdate('no action')->onDelete('no action');
+                $table->foreign(['usuario_id_usuario'], 'fk_reserva_usuario')->references(['id_usuario'])->on('usuario')->onUpdate('no action')->onDelete('no action');
+            });
+        }
     }
 
     /**
